requirements.txt
aiohttp @ file:///C:/ci/aiohttp_1614361024229/work
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
azure-core==1.14.0
azure-storage-blob==12.8.1
bokeh @ file:///C:/ci/bokeh_1620784067744/work
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247279197/work
chardet @ file:///C:/ci/chardet_1605303225733/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
click-plugins==1.1.1
cligj==0.7.1
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cryptography @ file:///C:/ci/cryptography_1616769344312/work
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1611341281313/work
distributed @ file:///C:/ci/distributed_1617384289923/work
Fiona==1.8.13.post1
fsspec @ file:///tmp/build/80754af9/fsspec_1617959894824/work
GDAL==3.0.2
geopandas==0.6.1
HeapDict==1.0.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
isodate==0.6.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
locket==0.2.1
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack @ file:///C:/ci/msgpack-python_1612287368835/work
msrest==0.6.21
multidict @ file:///C:/ci/multidict_1607362065515/work
munch==2.5.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile==0.46
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas @ file:///C:/ci/pandas_1618365634936/work
partd @ file:///tmp/build/80754af9/partd_1618000087440/work
Pillow @ file:///C:/ci/pillow_1617386341487/work
psutil @ file:///C:/ci/psutil_1612298324802/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyproj @ file:///C:/ci/pyproj_1614278136892/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
Rtree @ file:///C:/ci/rtree_1618421009405/work
Shapely==1.7.1
six @ file:///C:/ci/six_1605187374963/work
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1606865132123/work
tblib @ file:///tmp/build/80754af9/tblib_1597928476713/work
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
tornado @ file:///C:/ci/tornado_1606942392901/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
yarl @ file:///C:/ci/yarl_1606940076464/work
zict==2.0.0
